Coursework 3: The Depth Of A Surface
The task was to explore the Surface Generating Tool and strategies presented in the Course Session3. I decided to try on this coursework to solve the task from the Advent of Code Day 3 (https://adventofcode.com/2019/day/3) using Python and RhinoScriptSyntax. As an additional task (to make it more interesting) I decided to extrude Curves by the shortest distance from the beginning of coordinates to the closest point of intersection of the curves.